--- server/utils/routeHelpers.js
@@ -1,25 +1,0 @@
-// @flow
-
-export function githubUrl(): string {
-  return 'https://www.github.com/outline';
-}
-
-export function blogUrl(): string {
-  return 'https://medium.com/getoutline';
-}
-
-export function twitterUrl(): string {
-  return 'https://twitter.com/getoutline';
-}
-
-export function spectrumUrl(): string {
-  return 'https://spectrum.chat/outline';
-}
-
-export function developers(): string {
-  return '/developers';
-}
-
-export function signin(): string {
-  return '/auth/slack';
-}
