+++ shared/utils/routeHelpers.js
@@ -24,3 +24,27 @@ export function slackAuth(
 
   return `${baseUrl}?${urlParams}`;
 }
+
+export function githubUrl(): string {
+  return 'https://www.github.com/outline';
+}
+
+export function blogUrl(): string {
+  return 'https://medium.com/getoutline';
+}
+
+export function twitterUrl(): string {
+  return 'https://twitter.com/getoutline';
+}
+
+export function spectrumUrl(): string {
+  return 'https://spectrum.chat/outline';
+}
+
+export function developers(): string {
+  return '/developers';
+}
+
+export function signin(): string {
+  return '/auth/slack';
+}
